streamlit/pyproject.toml
[tool.poetry]
name = "streamlit_config"
version = "0.1.0"
description = ""
authors = ["dhirthacker7neu <thacker.dh@northeastern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.40.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
openai = "^1.56.1"
pinecone-client = "^5.0.1"
tiktoken = "^0.8.0"
scikit-learn = "^1.5.2"
google-api-python-client = "^2.154.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
